CI: WASM Pyodide (#1903)
* CI: WASM Pyodide job — build the wheel and run the Python test suite

Add a WASM CI job that cross-compiles the openPMD-api Python wheel for Pyodide
(openPMD_USE_PYTHON=ON) and runs the full Python unittest suite (APITest, incl.
the HDF5 read/write path) in the Pyodide test runtime via cibuildwheel's own
test runner -- so the test ABI matches the freshly built wheel. Catches
wasm-specific regressions a wheel smoke test cannot (e.g. the HDF5 64-bit long
double read path, #1902).

.github/ci/wasm_deps.sh cross-builds static zlib + HDF5 into the Emscripten
sysroot (Emscripten restricts find_package to the sysroot) incl. the HDF5
FE_INVALID emscripten patch. run_python_tests.py sets cwd + sys.path so the
suite's relative sample paths and API.APITest import resolve.


* WASM CI: add a co-load HDF5 teardown probe (openpmd_api + h5py)

Reproduce the ImpactX WASM teardown fault in openPMD-api's own CI: h5py bundles
a second static HDF5, which under Pyodide's single namespace interposes with
openpmd_api's HDF5. At interpreter teardown the tangled type registry makes the
openPMD HDF5 handler-destructor's H5Tclose fault ("not a datatype" -> wasm OOB),
even though the round-trip succeeds.

The probe runs after the full Python suite and is guarded with `|| echo` so a
crash is logged without failing the job while we reproduce (step 1). The guard
is dropped once the openPMD-api teardown fix lands, turning it into an enforced
regression test.


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* WASM CI: instrument the co-load probe (phase markers + HDF5 versions)

Run coload_repro.py first and standalone -- openpmd_api imported and run before
h5py loads (the ImpactX order, openPMD as the primary HDF5) -- printing a marker
per phase and both bundled HDF5 versions. This localizes any fatal fault
(mid-operation vs. teardown) and distinguishes an ABI mismatch (h5py bundling a
different HDF5) from the same-version teardown fault.

Pin the import order with `# isort: skip_file` so pre-commit's isort does not
alphabetize h5py ahead of openpmd_api (which would invert primary/secondary).
Probe and suite are both guarded (`|| echo`, `;`) so both run and the job stays
green while we diagnose from the log.


* WASM: keep bundled HDF5 private (-sSIDE_MODULE=2) to fix co-load crash

Pyodide links the Python extension as a side module with -sSIDE_MODULE=1, which
whole-archives and force-exports every symbol -- including the statically linked
HDF5. A second co-loaded wheel that also bundles HDF5 (ImpactX, h5py, a second
openPMD) then cross-binds its HDF5 symbols into ours, so two HDF5 instances share
one tangled registry and crash on the first file op or at teardown ("not a
datatype" -> wasm memory access out of bounds). -fvisibility=hidden cannot
prevent this because the whole-archive export overrides it.

Mirror the APPLE `-exported_symbol` / Linux `--exclude-libs` co-load fix on
Emscripten: -sSIDE_MODULE=2 exports only the symbols in EXPORTED_FUNCTIONS (just
the module init _PyInit_openpmd_api_cxx), keeping HDF5 private to this wheel.


* WASM: keep bundled HDF5 private via -sSIDE_MODULE=2 + hidden-visibility deps

The co-load crash (two wheels each embedding HDF5 in Pyodide's single namespace)
needs BOTH halves, neither of which suffices alone:

  - -sSIDE_MODULE=2 on the extension: Pyodide's default -sSIDE_MODULE=1 whole-
    archives and force-exports every symbol (so -fvisibility=hidden alone cannot
    hide HDF5); =2 exports only _PyInit_openpmd_api_cxx.
  - -fvisibility=hidden on the static zlib/HDF5 (wasm_deps.sh): makes their
    definitions DSO-local so wasm-ld can relax our GOT references to HDF5 into
    direct calls into our own copy, instead of binding to a co-loaded second
    HDF5 (h5py, ImpactX) at load time.

Without the visibility half, =2 stops the re-export but our HDF5 calls still bind
to the first-loaded global HDF5 via the GOT and crash on the first file op; this
was verified in CI (the co-load probe still OOB'd with =2 alone).


* WASM CI: enforce the co-load probe + fix the h5py-interop test dtype

Now that the co-load fix works, drop the `|| echo` guards so coload_repro.py and
the full suite are enforced regression tests: any wasm co-load regression fails
the job.

Also fix testScalarHdf5Fields, which ran for the first time on wasm (h5py is now
present in the test env): on wasm32 h5py writes the scalar with its native int
(int32) while np.array([45]) defaults to int64, so the store type-mismatched the
component ("store from Python array of type 'LONG' into Record Component of type
'INT'"). Store the component's own dtype instead of assuming the two agree.


* WASM CI: fix flake8 E501 (line too long) in the co-load test helpers

Wrap the over-79-character comment lines flagged by the `style` job:
coload_repro.py header, run_python_tests.py header, and the testScalarHdf5Fields
dtype comment in APITest.py. No code changes.


* WASM CI: relocate helpers + run Test.py directly (drop the wrapper)

- test/python/coload_repro.py (was .github/ci/): the co-load probe now lives
  with the other Python tests.
- .github/workflows/dependencies/install_wasm.sh (was .github/ci/wasm_deps.sh):
  next to where dependency installers belong.
- Drop run_python_tests.py: it only replicated cwd + sys.path[0]. Instead run
  `cd test/python/unittest && python Test.py -v`, mirroring the native ctest
  (CMakeLists.txt runs Test.py -v from that WORKING_DIRECTORY). Not pyodide-
  specific -- the ../samples relative paths require that cwd on every platform.


* Simplify: no need for inline cmd

* Improve inline comments

---------

+++ test/python/unittest/API/APITest.py
@@ -2333,7 +2333,7 @@ class APITest(unittest.TestCase):
 
         series_read_write = io.Series(file, io.Access.read_write)
         E_x = series_read_write.iterations[0].meshes["E"]["x"]
-        E_x[:] = np.array([45])
+        E_x[:] = np.array([45], dtype=E_x.dtype)
         series_read_write.close()
 
         series_read_again = io.Series(file, io.Access.read_only)
